+++ run_tests.sh
@@ -32,7 +32,7 @@ fi
 
 run() {
   # To update graalvm config, set GRAALVM_HOME, e.g:
-  # export GRAALVM_HOME=/usr/lib/jvm/java-21-graalvm
+  # export GRAALVM_HOME=/usr/lib/jvm/java-25-graalvm
   if [ ! -z "$GRAALVM_HOME" ]; then
     export JAVA_HOME=$GRAALVM_HOME
     export SIGNAL_CLI_OPTS="-agentlib:native-image-agent=config-merge-dir=graalvm-config-dir-${SIGNAL_CLI_AGENT_ID}/"